% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.R
\name{ABC}
\alias{ABC}
\title{ABC}
\usage{
ABC(data, na.rm = TRUE, plot = FALSE)
}
\arguments{
\item{data, }{Data frame of tuo columns,first column is the item name, second column is the item value/flow/demand.}

\item{na.rm, }{logical and by default is TRUE}

\item{plot, }{default is FALSE,if true a plot is generated}
}
\value{
a dataframe that contains ABC categories with a bar plot of the count of items in each category.
}
\description{
Identyfing ABC category based on the pareto rule.
Identyfing ABC category based on the pareto rule.A category is up to 80%. B category is up 95% and C category is up to 100%.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
ABC(data.frame(SKU= seq(1:1000),demand=runif(1000,1,1000)))
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
