% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_dynamic.R
\name{abc_dynamic}
\alias{abc_dynamic}
\title{abc_dynamic}
\usage{
abc_dynamic(
  product,
  key_to_split = F,
  first_attribute,
  second_attribute = F,
  A = F,
  B = F
)
}
\arguments{
\item{product, }{Vector that contains the product name .}

\item{key_to_split, }{logical and by default is False, otherwise a column that has a splitting dimension,
for example ; stores or cities}

\item{first_attribute}{, attribute to do the ABC analysis on, for example sales quantity}

\item{second_attribute}{, attribute to do the ABC analysis on .for example profit, the default is FALSE}

\item{A}{, changing the default threshold for A category which is 0.8, the default is FALSE}

\item{B}{, changing the default threshold for B category which is 0.95, the default is FALSE}
}
\value{
a dataframe that contains ABC categories.
}
\description{
Identyfing ABC category based on the pareto rule.
the function can have flexibility in defining the A,B thresholds. can be done on multiple splits for example
countries or stores
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
abc_dynamic(c(1:1000), rep(seq(1:10),100), runif(1000,4,10000),rnorm(1000,100,20))
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
