% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalRatio.R
\name{CriticalRatio}
\alias{CriticalRatio}
\title{Criticalratio}
\usage{
CriticalRatio(sellingprice, cost, salvage, penality, na.rm = TRUE)
}
\arguments{
\item{sellingprice}{numeric,selling price of the SKU}

\item{cost}{numeric,cost of the SKU}

\item{salvage}{numeric,,salvage or discounted value if sold after season,if there is no salvage , zero is placed in the argument.}

\item{penality}{numeric, peanlity cost of not satisfying demand if any, if not, zero is placed in the argument.}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
the critical ratio.
}
\description{
Calculating critical ratio of a news vendor model under any distribution.this critical ratio maxmizes profit.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
CriticalRatio(sellingprice=80,cost=60,salvage=45,penality=25,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
