% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSOE.R
\name{CSOE}
\alias{CSOE}
\title{CSOE}
\usage{
CSOE(quantity, demand, standerddeviation, leadtimeinweeks, cost, costSoe,
  holdingrate, na.rm = TRUE)
}
\arguments{
\item{quantity, }{numeric,quantity replinished every cycle.}

\item{demand}{numeric,annual Expected  demand of the SKU .}

\item{standerddeviation}{numeric,  standard  deviation of the SKU during season.}

\item{leadtimeinweeks}{numeric,leadtime in weeks of order.}

\item{cost}{numeric,cost of item.}

\item{costSoe}{numeric, estimated cost per stockout event.}

\item{holdingrate}{numeric, holding rate per item per year,percentage.}

\item{na.rm}{removes na values if TRUE, TRUE by default}
}
\value{
a dataframe that contains calculations of K the  item fill rate metric.cycle service level and expected unit short.
}
\description{
CSOE<- function(quantity,demand,standerddeviation,leadtimeinweeks,cost,costSoe,holdingrate,na.rm=TRUE)
}
\details{
Calculating K value that corresponds to the desired item fill rate.
}
\note{
this is the first version of the inventorize package, all the fucntions are common knowlege for supply chain without
any contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
CSOE(quantity=1000,demand=4000,standerddeviation=200,leadtimeinweeks=3,
cost=500,costSoe=85,holdingrate=0.2,na.rm=TRUE)
}
\author{
"haytham omar  email: <h.omar5942@gmail.com>"
}
