% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{elasticity}
\alias{elasticity}
\title{elasticity}
\usage{
elasticity(salesP1, salesP2, priceP1, priceP2)
}
\arguments{
\item{salesP1, }{integer, unit sales in period 1.}

\item{salesP2}{integer unit sales in period 2.}

\item{priceP1}{numeric,  average price of sku in period 1.}

\item{priceP2}{average price of sku in period 2.}
}
\value{
the elasticity ratio in unit sales, the -ve number represents the increase in sales for each decrease of unit currency.
}
\description{
calculating elasticity of price change.
}
\details{
This function is helpful to determine the elasticity of a product with effect to price change, the figure could be negative as the change is price is negative.
it translates as for each unit percentage  decrease in price , this much is ecpected precentage of  increase of sales. condition must be that Price in period one was more than proce
in period 2 and sales in period two was more than sales in period 1.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
elasticity(salesP1=50,salesP2=100,priceP1=6,priceP2=4)
}
\author{
"haytham omar  email: "<haytham@rescaleanalytics.com>"
}
