% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoq.R
\name{eoq}
\alias{eoq}
\title{eoq}
\usage{
eoq(annualdemand, orderingcost, purchasecost, holdingrate, na.rm = TRUE)
}
\arguments{
\item{annualdemand}{numeric,annual demand of the SKU.}

\item{orderingcost, }{numeric ordeing cost of the SKU}

\item{purchasecost}{,numeric, purchase cost per item}

\item{holdingrate}{numeric holding rate per item per year.}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
the eoq,cycle stock time in years and cycle stock time in weeks.
}
\description{
economic order quantity.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
eoq(annualdemand=5000,orderingcost=400,purchasecost=140,holdingrate=0.2,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
