% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoqsenstivity.R
\name{eoqsenstivity}
\alias{eoqsenstivity}
\title{eoqsenstivity}
\usage{
eoqsenstivity(quantity, quantityoptimal, na.rm = TRUE)
}
\arguments{
\item{quantity}{numeric,quantity ordered every order cycle.}

\item{quantityoptimal}{, numeric optimal quantity based on EOQ.}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
the rate of increase of total relevant cost compared to the EOQ.
}
\description{
the rate of increase of total relevant cost compared to the EOQ.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
eoqsenstivity(quantity=5400,quantityoptimal=6000,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
