% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPN_singleperiod.R
\name{EPN_singleperiod}
\alias{EPN_singleperiod}
\title{EPN_singleperiod}
\usage{
EPN_singleperiod(quantity, mean, standerddeviation, p, c, g, b,
  na.rm = TRUE)
}
\arguments{
\item{quantity, }{numeric,quantity replinished every cycle.}

\item{mean}{numeric,Expected  demand of the SKU during season.}

\item{standerddeviation}{numeric,  standard  deviation of the SKU during season.}

\item{p}{numeric,selling price of the SKU}

\item{c}{numeric,cost of the SKU}

\item{g}{numeric,,salvage or discounted value if sold after season,if there is no salvage , zero is placed in the argument.}

\item{b}{numeric, peanlity cost of not satisfying demand if any, if not, zero is placed in the argument.}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
a dataframe that contains calculations of the expected profit from a newsvendor model based on normal distribution.
}
\description{
calculating expected profit for a newsvendor model.
}
\details{
calculating expected profit for a newsvendor model. based on assumed normal distribution demand.
}
\note{
this is the first version of the inventorize package, all the fucntions are basic knowlege for supply chain without
any contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
EPN_singleperiod(quantity=40149,mean= 32000,standerddeviation= 11000,p=24,c=10.9,g=7,b=0,na.rm=TRUE)
}
\author{
"haytham omar  email: <h.omar5942@gmail.com>"
}
