% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventorize.R
\docType{package}
\name{inventorize}
\alias{inventorize}
\alias{inventorize-package}
\title{inventorize: Inventory Analytics And Cost Calculations.}
\description{
inventory analytics and cost calculations for SKUs.
}
\author{
\strong{Maintainer}: Haytham Omar \email{haytham@rescaleanalytics.com}

}
