% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventorymetricsCSL.R
\name{inventorymetricsCSL}
\alias{inventorymetricsCSL}
\title{inventorymetricsCSL}
\usage{
inventorymetricsCSL(
  csl,
  demand,
  standerddeviation,
  quantity,
  leadtime,
  cost,
  holdingrate,
  na.rm = TRUE
)
}
\arguments{
\item{csl}{numeric,required times of demand that is fullfilled from cycle stock}

\item{demand}{numeric,annual demand of the SKU.}

\item{standerddeviation}{numeric, annual standard  deviation}

\item{quantity, }{numeric,quantity replinished every cycle.}

\item{leadtime, }{numeric,leadtime in weeks}

\item{cost, }{numeric,cost of the SKU.}

\item{holdingrate}{numeric, holding rate per item per year.}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
a dataframe that contains
demand leadtime, sigmadl(standard deviation in leadtime), saftey factor k determined
based on item fillrate provided, unit normal loss function, expected units to be short, cycle service level, fill rate,implied cost
per stockout event, saftey stock and suggested reorder point.
}
\description{
calculating inventory metrics based on CYCLE SERVICE LEVEL.
}
\details{
cycle service level is the desired no of times demand is compleltey fulfiiled from cycle stock,after cycle service level  is explicitly  calculated, cost per item short, cost per stock out event and item fill rate
are implicitly calculated.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
inventorymetricsCSL(csl=0.95,demand=20000,standerddeviation=1200,
quantity=4500,leadtime=3,cost=100,holdingrate=0.15,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
