% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventorymetricsIFR.R
\name{inventorymetricsIFR}
\alias{inventorymetricsIFR}
\title{inventorymetricsIFR}
\usage{
inventorymetricsIFR(fillrate, demand, standerddeviation, quantity,
  leadtime, cost, holdingrate, na.rm = TRUE)
}
\arguments{
\item{fillrate}{numeric,required percentage of demand that is fullfilled from cycle stock}

\item{demand}{numeric,annual demand of the SKU.}

\item{standerddeviation}{numeric, annual standard  deviation}

\item{quantity, }{numeric,quantity replinished every cycle.}

\item{leadtime, }{numeric,leadtime in weeks}

\item{cost, }{numeric cost of the SKU}

\item{holdingrate}{,numeric, holding rate per item/year}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
a dataframe that contains
demand leadtime, sigmadl(standard deviation in leadtime), saftey factor k determined
based on item fillrate provided, unit normal loss function expected units to be short,cycle service level, fill rate,implied cost
per stockout event, saftey stock and suggested reorder point.
}
\description{
calculating inventory metrics based on item fillrate.
}
\details{
item fill rate is the percentage of demand that is fullfilled directly from the cycle stock,after item fill rate is explicitly  calculated, cost per item short, cost per stock out event and cycle service level
are implicitly calculated.
}
\note{
this is the first version of the inventorize package, all the fucntions are common knowlege for supply chain without
any contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during analysis of stock.
the package relies heavily on the leraning from MIT micromasters.
}
\examples{
inventorymetricsIFR(fillrate= 0.90, demand= 35000,standerddeviation=9000,
quantity= 5000,leadtime=3 ,cost=50,holdingrate=0.15,na.rm=TRUE)
}
\author{
"haytham omar  email: <h.omar5942@gmail.com>"
}
