% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPN_singleperiod.R
\name{MPN_singleperiod}
\alias{MPN_singleperiod}
\title{MPN_singleperiod}
\usage{
MPN_singleperiod(mean, standerddeviation, p, c, g, b, na.rm = TRUE)
}
\arguments{
\item{mean}{numeric,Expected  demand of the SKU during season.}

\item{standerddeviation}{numeric,  standard  deviation of the SKU during season.}

\item{p}{numeric,selling price of the SKU}

\item{c}{numeric,cost of the SKU}

\item{g}{numeric,,salvage or discounted value if sold after season,if there is no salvage , zero is placed in the argument.}

\item{b}{numeric, peanlity cost of not satisfying demand if any, if not, zero is placed in the argument.}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
a dataframe that contains calculations of the maximum expected profit from a newsvendor model based on normal distribution.
}
\description{
calculating expected profit for a newsvendor model based on critical ratio.
}
\details{
calculating expected profit for a newsvendor model. based on assumed normal distribution demand.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
MPN_singleperiod(mean= 32000,standerddeviation= 11000,p=24,c=10.9,g=7,b=0,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
