% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodic_review_pois.R
\name{Periodic_review_pois}
\alias{Periodic_review_pois}
\title{Periodic_review_pois}
\usage{
Periodic_review_pois(
  demand,
  lambda,
  leadtime,
  service_level,
  Review_period,
  shortage_cost = FALSE,
  inventory_cost = FALSE,
  ordering_cost = FALSE
)
}
\arguments{
\item{demand}{A vector of demand in N time periods.}

\item{lambda}{rate of demand in N time periods.}

\item{leadtime}{lead time from order to arrival}

\item{service_level}{cycle service level requested}

\item{Review_period}{the period where the ordering happens.}

\item{shortage_cost}{shortage cost per unit of sales lost}

\item{inventory_cost}{inventory cost per unit.}

\item{ordering_cost}{ordering cost for every time an order is made.}
}
\value{
a list of two date frames, the simulation and the metrics.
}
\description{
Simulating a Periodic order up to level policy,  .
}
\details{
The Function takes a demand vector, mean of demand ,sd,lead time and 
requested service level to simulate and inventory system, 
orders are lost if inventory level is less than requested demand, also ordering is made at
 day t+1, metrics like item fill rate and cycle service level are calculated.
  the order up to level is calculated based on the review period,lead time and Poisson distribution .
}
\examples{
Periodic_review_pois(demand=rpois(80,6),lambda=6,leadtime=5,service_level=0.95,
Review_period =9,
shortage_cost= FALSE,inventory_cost=FALSE,ordering_cost=FALSE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
