% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_mix.R
\name{productmix}
\alias{productmix}
\title{productmix}
\usage{
productmix(SKUs, sales, revenue, na.rm = TRUE)
}
\arguments{
\item{SKUs, }{charachter, a vector of SKU names.}

\item{sales, }{vector, a vector of items sold per sku, should be the same number of rows as SKU.}

\item{revenue}{price  vector, a vector of total revenu  per sku, should be the same number of rows as SKU.}

\item{na.rm}{, logical and by default is TRUE}
}
\value{
a dataframe that contains ABC categories with a bar plot of the count of items in each category.
}
\description{
Identyfing ABC category based on the pareto rule for both demand and selling price,a mix of nine categories are produced.
Identyfing ABC category based on the pareto rule.A category is up to 80%. B category is up 95% and C category is up to 100%.
}
\note{
this is the first version of the inventorize package, all the fucntions are common knowlege for supply chain without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
productmix(SKUs=c(1:100),sales=runif(100,1,1000),revenue = rnorm(100,200,10),na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
