% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderpoint_normal.R
\name{reorderpoint}
\alias{reorderpoint}
\title{reorderpoint}
\usage{
reorderpoint(dailydemand, dailystandarddeviation, leadtimein_days, csl,
  na.rm = TRUE)
}
\arguments{
\item{dailydemand}{numeric,daily Expected  demand of the SKU .}

\item{dailystandarddeviation}{numeric,  standard  deviation of daily demand of the SKU .}

\item{leadtimein_days}{leadtime in days of order..}

\item{csl}{cycle service level requested}

\item{na.rm}{Logical, remove na if TRUE}
}
\value{
a dataframe that contains demand lead time,sigmadl,safteyfactor and re_order point.
}
\description{
Calculating saftey stock based on the cycle service level.
}
\details{
Calculating re-order point  based on  demand variability without lead time variability in   an assumed normal distribution. cycle service level is provided to calculate saftey stock accordingly.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
reorderpoint(dailydemand=50,dailystandarddeviation=5,leadtimein_days=6,csl=0.90)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
