% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_max.R
\name{revenue_max}
\alias{revenue_max}
\title{revenue_max}
\usage{
revenue_max(salesP1, salesP2, priceP1, priceP2, na.rm = TRUE)
}
\arguments{
\item{salesP1, }{integer, unit sales in period 1.}

\item{salesP2}{integer unit sales in period 2.}

\item{priceP1}{numeric,  average price of sku in period 1.}

\item{priceP2}{average price of sku in period 2.}

\item{na.rm}{logical with a default of TRUE}
}
\value{
the elasticity ratio in unit sales, the -ve number represents the increase in sales for each decrease of unit currency.
}
\description{
maxmizing revenue based on chage in price and elasticity.
}
\details{
#' This function is helpful to determine the elasticity of a product with effect to price change, the figure could be negative as the change is price is negative.
it translates as for each unit percentage  decrease in price , this much is ecpected precentage of  increase of sales. condition must be that Price in period one was more than proce
in period 2 and sales in period two was more than sales in period 1.  a proposed optimum  price is given to period 3 which is future period to maxmize revenue.
}
\examples{
revenue_max(salesP1=50,salesP2=100,priceP1=6,priceP2=4)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
