% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saftey_stock_normal.R
\name{saftey_stock_normal}
\alias{saftey_stock_normal}
\title{saftey_stock_normal}
\usage{
saftey_stock_normal(annualdemand, annualstandarddeviation, leadtimeinweeks,
  csl, na.rm = TRUE)
}
\arguments{
\item{annualdemand}{numeric,annual Expected  demand of the SKU .}

\item{annualstandarddeviation}{numeric,  standard  deviation of the SKU during season.}

\item{leadtimeinweeks}{leadtime in weeks or order.}

\item{csl}{cycle service level requested}

\item{na.rm}{Logical, remove na if TRUE}
}
\value{
a dataframe that contains calculations of K the cost per item short metric noting that condition must me less than 1.
}
\description{
Calculating saftey stock based on the cycle service level.
}
\details{
Calculating saftey stock  based on  the cycle service level in   an assumed normal distribution.
}
\note{
this is the first version of the inventorize package, all the fucntions are basic knowlege for supply chain without
any contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
saftey_stock_normal(annualdemand=8000,annualstandarddeviation=600,
leadtimeinweeks=4,csl=0.92,na.rm=TRUE)
}
\author{
"haytham omar  email: <h.omar5942@gmail.com>"
}
