% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safteystock_IFR_normal.R
\name{safteystock_IFR_normal}
\alias{safteystock_IFR_normal}
\title{safteystock_IFR_normal}
\usage{
safteystock_IFR_normal(rate, quantity, demand, standerddeviation, leadtime,
  na.rm = TRUE)
}
\arguments{
\item{rate}{numeric, item fill rate.}

\item{quantity, }{numeric,quantity replinished every cycle.}

\item{demand}{numeric,annual Expected  demand of the SKU .}

\item{standerddeviation}{numeric,  standard  deviation of the SKU during season.}

\item{leadtime}{leadtime in weeks of order.}

\item{na.rm}{Logical, TRUE to remove na.}
}
\value{
a dataframe that contains calculations of K the  item fill rate metric.cycle service level and expected unit short.
}
\description{
Calculating K value corresponding to item fill rate.
}
\details{
Calculating K value that corresponds to the desired item fill rate.
}
\note{
this is the first version of the inventorize package, all the fucntions are basic knowlege for supply chain without
any contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
safteystock_IFR_normal(rate=0.97,quantity=9000,demand=100000,
standerddeviation=5000,leadtime=4,na.rm=TRUE)
}
\author{
"haytham omar  email: <h.omar5942@gmail.com>"
}
