% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.logistics.cost.R
\name{total.logistics.cost}
\alias{total.logistics.cost}
\title{total.logistics.cost}
\usage{
total.logistics.cost(
  quantity,
  expected_annual_demand,
  sd_annual_demand,
  expected_leadtimeindays,
  sd_leadtime,
  costperunit,
  transportcost,
  holdingrate,
  ordering_cost,
  csl
)
}
\arguments{
\item{quantity}{quantity ordered every cycle.}

\item{expected_annual_demand}{numeric,  expected annual demand of the SKU.}

\item{sd_annual_demand}{annual standard deviation of the SKU.}

\item{expected_leadtimeindays}{expected lead time in days.}

\item{sd_leadtime}{standard deviation of leadtime}

\item{costperunit}{purchase cost of the SKU}

\item{transportcost}{transport cost of the SKU}

\item{holdingrate}{holding rate of the SKU}

\item{ordering_cost}{ordering cost per order placed}

\item{csl}{cycle service level desired}
}
\value{
a dataframe that contains calculations of the total logistics cost in detail.
}
\description{
calculating total logistics cost .
}
\details{
calculating total logistics cost based on a normal distribution.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
total.logistics.cost(quantity=32,expected_annual_demand=1550,
sd_annual_demand=110,expected_leadtimeindays=64,sd_leadtime=8,
costperunit=107,transportcost=22,holdingrate=0.15,ordering_cost=500,csl=0.95)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
