% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TQpractical.R
\name{TQpractical}
\alias{TQpractical}
\title{TQpractical}
\usage{
TQpractical(annualdemand, orderingcost, purchasecost, holdingrate,
  na.rm = TRUE)
}
\arguments{
\item{annualdemand, }{numeric annual demand of the SKU.}

\item{orderingcost, }{numeric  ordering cost of the SKU.}

\item{purchasecost}{numeric purchase cost of the SKU.}

\item{holdingrate}{numeric holding rate of the SKU.}

\item{na.rm}{logical, TRUE.}
}
\value{
a dataframe that contains the economic order quantity and the practical order quantity, Tstar (optimum)and Tpractical
which is always away from the optimum up to 6%.
}
\description{
Identyfing Practical ordering quantity based on the economic order quantity.it is assumed that practical
order quantity will be always withing 6 % of the economic order quantity in terms od total relevant cost.
}
\note{
this is the first version of the inventorize package, all the functions are common knowlege for supply chain without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
the package heavily relies on the learnings from MIT micromasters in supply chain led by Prof. Chris Caplice.
}
\examples{
TQpractical(annualdemand=1000,orderingcost=100,purchasecost=72,holdingrate=0.25,na.rm=TRUE)
}
\author{
"haytham omar  email: <h.omar5942@gmail.com>"
}
