% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRC.R
\name{TRC}
\alias{TRC}
\title{TRC}
\usage{
TRC(annualdemand, orderingcost, purchasecost, holdingrate, na.rm = TRUE)
}
\arguments{
\item{annualdemand}{numeric annual demand of the SKU.}

\item{orderingcost}{numeric  ordering cost of the SKU.}

\item{purchasecost}{numeric purchase cost of the SKU.}

\item{holdingrate}{numeric holding rate of the SKU.}

\item{na.rm}{logical, TRUE to remove na.}
}
\description{
Identyfing Total relevant cost.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
TRC(annualdemand=2500,orderingcost=250,purchasecost=98,holdingrate=0.25,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
