% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSOE.R
\name{CSOE}
\alias{CSOE}
\title{CSOE}
\usage{
CSOE(quantity, demand, standerddeviation, leadtimeinweeks, cost, costSoe,
  holdingrate, na.rm = TRUE)
}
\arguments{
\item{quantity, }{numeric,quantity replinished every cycle.}

\item{demand}{numeric,annual Expected  demand of the SKU .}

\item{standerddeviation}{numeric,  standard  deviation of the SKU during season.}

\item{leadtimeinweeks}{numeric,leadtime in weeks of order.}

\item{cost}{numeric,cost of item.}

\item{costSoe}{numeric, estimated cost per stockout event.}

\item{holdingrate}{numeric, holding rate per item per year,percentage.}

\item{na.rm}{removes na values if TRUE, TRUE by default}
}
\value{
a dataframe that contains calculations of K and the minimum quantity to be put in stock .
}
\description{
Cost per stockout event
}
\details{
Calculating K value that corresponds to the cost per stock out event, how much quantity should be put in stock as a minimum.the function solves for optimum K
based on the stock out event. It should be noted that the condition(output) should be bigger than 1.
other wise set K as per management.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
CSOE(quantity=1000,demand=40000,standerddeviation=200,leadtimeinweeks=3,
cost=500,costSoe=30000,holdingrate=0.2,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
