% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPP_singleperiod.R
\name{EPP_singleperiod}
\alias{EPP_singleperiod}
\title{EPP_singleperiod}
\usage{
EPP_singleperiod(quantity, lambda, p, c, g, b, na.rm = TRUE)
}
\arguments{
\item{quantity}{numeric,quantity to be ordered during season.}

\item{lambda}{numeric,  mean of the demand based on poisson distribution.}

\item{p}{numeric,selling price of the SKU}

\item{c}{numeric,cost of the SKU}

\item{g}{numeric,,salvage or discounted value if sold after season,if there is no salvage , zero is placed in the argument.}

\item{b}{numeric, peanlity cost of not satisfying demand if any, if not, zero is placed in the argument.}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
a dataframe that contains calculations of the  expected profit from a newsvendor model based on poisson distribution.
}
\description{
Expected profit from a newsvendor model based on a poisson distribution.
}
\details{
calculating expected profit for a newsvendor model. based on assumed poisson distribution demand.
}
\note{
this is the second version of the inventorize package, all the fucntions are  without
any academic contribution from my side, the aim is to facilitate and ease much of the bookkeeping that is endured during stock analysis.
}
\examples{
EPP_singleperiod(quantity=40149,lambda= 32000,p=24,c=10.9,g=7,b=0,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
