% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safteystock_CSL_normal.R
\name{safteystock_CSL_normal}
\alias{safteystock_CSL_normal}
\title{safteystock_CSL_normal}
\usage{
safteystock_CSL_normal(rate, quantity, demand, standerddeviation, leadtime,
  na.rm = TRUE)
}
\arguments{
\item{rate, }{cycle service level requested.}

\item{quantity}{quantity ordered every cycle.}

\item{demand}{numeric,  expected annual demand of the SKU.}

\item{standerddeviation}{numeric annual standard deviation of the demand.}

\item{leadtime}{numeric,leadtime of order in weeks.}

\item{na.rm}{logical with a default of TRUE}
}
\value{
a dataframe that contains calculations of the expected profit from a newsvendor model based on normal distribution.
}
\description{
calculating saftey stock based on cycle service level rate.
}
\details{
calculating saftey stock and expected unit short based on the cycle service identified assuming a normal distribution.
}
\examples{
safteystock_CSL_normal(rate=0.95,quantity=30000,demand=28000,standerddeviation=5000,8,na.rm=TRUE)
}
\author{
"haytham omar  email: <haytham@rescaleanalytics.com>"
}
